[KED-1657] Seed randomly-generated data with UUIDs (#178)
- Generate a random UUID, and log it to the console, if random data is selected. Use it to seed all the random number generators.
- Allow users to copy the ID from the console and add it to a query string in the URL, so that the randomly-generated data can be replicated.
