+++ package.json
@@ -48,6 +48,7 @@
     "react-test-renderer": "^16.8.6",
     "redux": "^4.0.1",
     "reselect": "^4.0.0",
+    "seedrandom": "^3.0.5",
     "snyk": "^1.316.1",
     "svg-crowbar": "^0.3.1",
     "what-input": "^5.2.3"
