+++ trufflehog-ignore.txt
@@ -2,3 +2,4 @@ package.json
 package-lock.json
 RELEASE.md
 README.md
+src/utils/random-utils.js
